feat: rename VscodeTableInternal to VscodeTableInternalFix
