+++ packages/vscode-extension/src/network/components/VscodeTableInternalFix.tsx
